corelib + font demo 	+ fixed fltk Win32 DLL compilation problem because 		Mark class lacked FL_API decl. in Browser.h 	+ fixed several warnings because of incorrect type and casts of (float) font sizes.
git-svn-id: file:///fltk/svn/trunk@5516 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/drawtext.cxx
@@ -132,7 +132,7 @@ public:
       n = getsize()+n;
     else if (n <= 0)
       n = normal_size;
-    setfont(getfont(), n);
+    setfont(getfont(), (float) n);
   }
   void _draw(const Rectangle&) const {dostuff();}
   void _measure(int& w, int& h) const {dostuff(); w = 0;}
